Add ability to monitor progress of an individual chunk.
This makes multipart uploads appear to be progressing at the actual rate
that they are.

Includes a BUGWATCH, if we update the library to do concurrent uploads
of multiple chunks the math to calculate progress will have to change.

xhr-mock has some major issues that are fixed in their major v2, but are
not available properly on npm yet.